+++ jmp-application/pom.xml
@@ -11,4 +11,17 @@
 
 	<artifactId>jmp-application</artifactId>
 
+	<dependencies>
+		<dependency>
+			<groupId>com.epam</groupId>
+			<artifactId>jmp-bank-api</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+		</dependency>
+	</dependencies>
+
 </project>
