--- jmp-bank-api/src/main/java/com/epam/engx/jmp/api/Bank.java
@@ -1,9 +1,0 @@
-package com.epam.engx.jmp.api;
-
-import com.epam.engx.jmp.dto.BankCard;
-import com.epam.engx.jmp.dto.BankCardType;
-import com.epam.engx.jmp.dto.User;
-
-public interface Bank {
-	BankCard createBankCard(User user, BankCardType type);
-}
