--- jmp-bank-api/src/main/java/com/epam/engx/jmp/api/Service.java
@@ -1,16 +1,0 @@
-package com.epam.engx.jmp.api;
-
-import java.util.List;
-import java.util.Optional;
-
-import com.epam.engx.jmp.dto.BankCard;
-import com.epam.engx.jmp.dto.Subscription;
-import com.epam.engx.jmp.dto.User;
-
-public interface Service {
-	void subscribe(BankCard bankCard);
-
-	Optional<Subscription> getSubscriptionByBankCardNumber(String cardNumber);
-
-	List<User> getAllUsers();
-}
