+++ jmp-bank-api/src/main/java/com/epam/jmp/api/Bank.java
@@ -1,0 +1,9 @@
+package com.epam.jmp.api;
+
+import com.epam.jmp.dto.BankCard;
+import com.epam.jmp.dto.BankCardType;
+import com.epam.jmp.dto.User;
+
+public interface Bank {
+	BankCard createBankCard(User user, BankCardType type);
+}
