+++ jmp-bank-api/src/main/java/module-info.java
@@ -1,4 +1,4 @@
 module jmp.bank.api {
 	requires jmp.dto;
-	exports com.epam.engx.jmp.api;
+	exports com.epam.jmp.api;
 }
