+++ jmp-cloud-bank-impl/src/main/java/com/epam/jmp/cloud/bank/cardnumber/LuhnValidator.java
@@ -1,0 +1,42 @@
+package com.epam.jmp.cloud.bank.cardnumber;
+
+import java.util.function.Predicate;
+import java.util.regex.Pattern;
+
+/**
+ * A class that provides Card Number validation functionality.
+ *
+ * <p>
+ * This class implements the {@link Predicate} interface, allowing it to be used for testing a card number string.
+ * The card number should not be null, consist of only digits (0-9), and follow the Luhn algorithm for validation.
+ * </p>
+ */
+public final class LuhnValidator implements Predicate<String> {
+	private static final Predicate<String> CARD_NUMBER_VALIDATOR = Pattern.compile("\\d{2,19}").asPredicate();
+
+	@Override
+	public boolean test(String cardNumber) {
+		if (cardNumber == null) {
+			return false;
+		}
+		var number = cardNumber.replace(" ", "");
+		if (!CARD_NUMBER_VALIDATOR.test(number)) {
+			return false;
+		}
+
+		var sum = 0;
+		var isEven = number.length() % 2 == 0;
+		for (var symbol : number.toCharArray()) {
+			var digit = Character.getNumericValue(symbol);
+			if (isEven) {
+				digit *= 2;
+				if (digit > 9) {
+					digit -= 9;
+				}
+			}
+			sum += digit;
+			isEven = !isEven;
+		}
+		return sum % 10 == 0;
+	}
+}
