+++ jmp-cloud-bank-impl/src/main/java/module-info.java
@@ -1,8 +1,9 @@
-import com.epam.engx.jmp.api.Bank;
-import com.epam.engx.jmp.cloud.bank.BankImpl;
+import com.epam.jmp.api.Bank;
+import com.epam.jmp.cloud.bank.BankImpl;
 
 module jmp.cloud.bank.impl {
 	requires transitive jmp.bank.api;
 	requires jmp.dto;
-	provides Bank with BankImpl;
+	provides Bank
+		with BankImpl;
 }
