+++ jmp-cloud-service-impl/src/main/java/com/epam/jmp/cloud/service/ServiceImpl.java
@@ -1,0 +1,28 @@
+package com.epam.jmp.cloud.service;
+
+
+import com.epam.jmp.dto.BankCard;
+import com.epam.jmp.dto.Subscription;
+import com.epam.jmp.dto.User;
+import com.epam.jmp.service.api.Service;
+
+import java.util.List;
+import java.util.Optional;
+
+public class ServiceImpl implements Service {
+
+	@Override
+	public void subscribe(BankCard bankCard) {
+
+	}
+
+	@Override
+	public Optional<Subscription> getSubscriptionByBankCardNumber(String cardNumber) {
+		return Optional.empty();
+	}
+
+	@Override
+	public List<User> getAllUsers() {
+		return null;
+	}
+}
