+++ jmp-cloud-service-impl/src/main/java/module-info.java
@@ -1,6 +1,9 @@
+import com.epam.jmp.cloud.service.ServiceImpl;
+import com.epam.jmp.service.api.Service;
+
 module jmp.cloud.service.impl {
 	requires transitive jmp.service.api;
 	requires jmp.dto;
-	provides com.epam.engx.jmp.service.api.Service
-		with com.epam.engx.jmp.cloud.service.ServiceImpl;
+	provides Service
+		with ServiceImpl;
 }
