--- jmp-dto/src/main/java/com/epam/engx/jmp/dto/BankCardType.java
@@ -1,5 +1,0 @@
-package com.epam.engx.jmp.dto;
-
-public enum BankCardType {
-	CREDIT, DEBIT
-}
