--- jmp-dto/src/main/java/com/epam/engx/jmp/dto/CreditBankCard.java
@@ -1,7 +1,0 @@
-package com.epam.engx.jmp.dto;
-
-public final class CreditBankCard extends BankCard {
-	public CreditBankCard(String number, User user) {
-		super(number, user);
-	}
-}
