--- jmp-dto/src/main/java/com/epam/engx/jmp/dto/DebitBankCard.java
@@ -1,7 +1,0 @@
-package com.epam.engx.jmp.dto;
-
-public final class DebitBankCard extends BankCard {
-	public DebitBankCard(String number, User user) {
-		super(number, user);
-	}
-}
