--- jmp-dto/src/main/java/com/epam/engx/jmp/dto/Subscription.java
@@ -1,11 +1,0 @@
-package com.epam.engx.jmp.dto;
-
-import java.time.LocalDate;
-import java.util.Objects;
-
-public record Subscription(String bankcard, LocalDate startDate) {
-	public Subscription {
-		Objects.requireNonNull(bankcard, "Bankcard cannot be null");
-		Objects.requireNonNull(startDate, "Start date cannot be null");
-	}
-}
