+++ jmp-dto/src/main/java/com/epam/jmp/Application.java
@@ -1,0 +1,10 @@
+package com.epam.jmp;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+	public static void main(String[] args) {
+		SpringApplication.run(Application.class, args);
+	}
+}
