+++ jmp-dto/src/main/java/com/epam/jmp/dto/BankCardType.java
@@ -1,0 +1,5 @@
+package com.epam.jmp.dto;
+
+public enum BankCardType {
+	CREDIT, DEBIT
+}
