+++ jmp-dto/src/main/java/com/epam/jmp/dto/CreditBankCard.java
@@ -1,0 +1,7 @@
+package com.epam.jmp.dto;
+
+public final class CreditBankCard extends BankCard {
+	public CreditBankCard(String number, User user) {
+		super(number, user);
+	}
+}
