+++ jmp-dto/src/main/java/com/epam/jmp/dto/DebitBankCard.java
@@ -1,0 +1,7 @@
+package com.epam.jmp.dto;
+
+public final class DebitBankCard extends BankCard {
+	public DebitBankCard(String number, User user) {
+		super(number, user);
+	}
+}
