+++ jmp-dto/src/main/java/com/epam/jmp/entity/User.java
@@ -1,0 +1,30 @@
+package com.epam.jmp.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Past;
+import lombok.Data;
+
+import java.time.LocalDate;
+
+@Data
+@Entity
+@Table(name = "bank_user")
+public class User {
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private Long id;
+
+	@NotBlank
+	private String name;
+	@NotBlank
+	private String surname;
+	@Past
+	@NotNull
+	private LocalDate birthday;
+}
