+++ jmp-dto/src/main/java/com/epam/jmp/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.epam.jmp.repository;
+
+import com.epam.jmp.entity.User;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+	Optional<User> findByName(String name);
+}
