+++ jmp-dto/src/main/java/module-info.java
@@ -1,4 +1,14 @@
-module jmp.dto {
-	exports com.epam.engx.jmp.dto;
+open module jmp.dto {
+	exports com.epam.jmp.dto;
+
+	requires lombok;
 	requires java.desktop;
+    requires jakarta.persistence;
+	requires jakarta.validation;
+	requires spring.data.jpa;
+	requires spring.context;
+	requires spring.beans;
+	requires spring.boot;
+    requires spring.boot.autoconfigure;
+	requires spring.web;
 }
