+++ jmp-service-api/src/main/java/module-info.java
@@ -1,4 +1,4 @@
 module jmp.service.api {
 	requires jmp.dto;
-	exports com.epam.engx.jmp.service.api;
+	exports com.epam.jmp.service.api;
 }
